Add initial group changing
