+++ app/src/main/java/ru/neexol/rtut/presentation/screens/main/ui/MainScreen.kt
@@ -11,20 +11,25 @@ import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.lifecycle.viewmodel.compose.viewModel
 import com.google.accompanist.pager.ExperimentalPagerApi
+import ru.neexol.rtut.presentation.screens.settings.SettingsViewModel
 
 @ExperimentalAnimationApi
 @ExperimentalMaterialApi
 @ExperimentalPagerApi
 @Composable
 fun MainScreen() {
-	val sheetState = rememberModalBottomSheetState(ModalBottomSheetValue.Hidden)
-	val isSheetHidden by remember {
-		derivedStateOf { !sheetState.isVisible }
-	}
+	val vm: SettingsViewModel = viewModel()
+	if (!vm.groupUiState.group.isNullOrEmpty()) {
+		val sheetState = rememberModalBottomSheetState(ModalBottomSheetValue.Hidden)
+		val isSheetHidden by remember {
+			derivedStateOf { !sheetState.isVisible }
+		}
 
-	Column(Modifier.navigationBarsPadding()) {
-		StatusBar(!isSheetHidden)
-		Screens(sheetState, isSheetHidden)
-	}
+		Column(Modifier.navigationBarsPadding()) {
+			StatusBar(!isSheetHidden)
+			Screens(sheetState, isSheetHidden, vm)
+		}
+	} else InitialScreen(vm.groupUiState.group?.run { vm::editGroup })
 }
