+++ app/src/main/java/ru/neexol/rtut/presentation/screens/settings/ui/SettingsScreen.kt
@@ -8,11 +8,10 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
 import ru.neexol.rtut.presentation.screens.settings.SettingsViewModel
 
 @Composable
-fun SettingsScreen(vm: SettingsViewModel = hiltViewModel()) {
+fun SettingsScreen(vm: SettingsViewModel) {
 	Column {
 		LogoBar()
 		Column(
