+++ app/src/main/res/values-ru-rRU/strings.xml
@@ -35,4 +35,5 @@
     <string name="copy">Копировать</string>
     <string name="restore">Восстановить</string>
     <string name="about_app" tools:ignore="TypographyDashes">Приложение написано в качестве ВКР, Алексеев Н.Е., ИКБО-12-18, РТУ МИРЭА, 2022.</string>
+    <string name="enter_group">Введите группу:</string>
 </resources>
